Move "helper" classes to existing package with similar content
